Reimplement lag plotter and add features
This is a complete reimplementation of the lag plotting script.
The main goal is to show the lag of each MPI process rather than
rank-0 only.  However, the previous implementation handled times
and plots in a somewhat overcomplicated and fragile way, so I did
a full rewrite to make my life a bit easier.

The new plots also show UTC times regardless of the timezone of
the machines running PyCBC Live, so we can compare plots across
different data centers easily, and also compare with the detchar
summary pages, to easily correlate lag issues with DQ issues.
The log->GPS conversion is not correct for all times, but the
old script had a very similar issue, so I decided to keep that
complication as a future fix.

Finally, the new code seems to be significantly faster (under a
minute per plot) although I have not done a careful comparison.